+++ crates/ruma-common/src/events/encrypted.rs
@@ -23,11 +23,7 @@ pub struct EncryptedEventContent {
     pub encrypted: EncryptedContentBlock,
 
     /// Information about related events.
-    #[serde(
-        flatten,
-        skip_serializing_if = "Option::is_none",
-        deserialize_with = "super::room::encrypted::relation_serde::deserialize_relation"
-    )]
+    #[serde(rename = "m.relates_to", skip_serializing_if = "Option::is_none")]
     pub relates_to: Option<Relation>,
 }