+++ crates/ruma-common/src/events/room/message/relation_serde.rs
@@ -1,7 +1,8 @@
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use serde::{de, Deserialize, Deserializer, Serialize};
+use serde_json::Value as JsonValue;
 
 use super::{InReplyTo, Relation, Replacement, Thread};
-use crate::OwnedEventId;
+use crate::{events::relation::CustomRelation, serde::JsonObject, OwnedEventId};
 
 /// Deserialize an event's `relates_to` field.
 ///
@@ -25,39 +26,39 @@ where
     D: Deserializer<'de>,
     C: Deserialize<'de>,
 {
-    let ev = EventWithRelatesToJsonRepr::deserialize(deserializer)?;
-
-    if let Some(
-        RelationJsonRepr::ThreadStable(ThreadStableJsonRepr { event_id, is_falling_back })
-        | RelationJsonRepr::ThreadUnstable(ThreadUnstableJsonRepr { event_id, is_falling_back }),
-    ) = ev.relates_to.relation
-    {
-        let in_reply_to = ev.relates_to.in_reply_to;
-        return Ok(Some(Relation::Thread(Thread { event_id, in_reply_to, is_falling_back })));
-    }
-
-    let rel = if let Some(in_reply_to) = ev.relates_to.in_reply_to {
-        Some(Relation::Reply { in_reply_to })
-    } else if let Some(relation) = ev.relates_to.relation {
-        match relation {
-            RelationJsonRepr::Replacement(ReplacementJsonRepr { event_id }) => {
-                let new_content = ev
-                    .new_content
-                    .ok_or_else(|| serde::de::Error::missing_field("m.new_content"))?;
-                Some(Relation::Replacement(Replacement { event_id, new_content }))
-            }
-            // FIXME: Maybe we should log this, though at this point we don't even have
-            // access to the rel_type of the unknown relation.
-            RelationJsonRepr::Unknown => Some(Relation::_Custom),
-            RelationJsonRepr::ThreadStable(_) | RelationJsonRepr::ThreadUnstable(_) => {
-                unreachable!()
-            }
-        }
-    } else {
-        None
+    let EventWithRelatesToDeHelper { relates_to, new_content } =
+        EventWithRelatesToDeHelper::deserialize(deserializer)?;
+    let Some(relates_to) = relates_to else {
+        return Ok(None);
     };
 
-    Ok(rel)
+    let RelatesToDeHelper { in_reply_to, relation } = relates_to;
+
+    let rel = if let Some(RelationDeHelper::Known(relation)) = relation {
+        match relation {
+            KnownRelationDeHelper::Replacement(ReplacementJsonRepr { event_id }) => {
+                match new_content {
+                    Some(new_content) => {
+                        Relation::Replacement(Replacement { event_id, new_content })
+                    }
+                    None => return Err(de::Error::missing_field("m.new_content")),
+                }
+            }
+            KnownRelationDeHelper::Thread(ThreadDeHelper { event_id, is_falling_back })
+            | KnownRelationDeHelper::ThreadUnstable(ThreadUnstableDeHelper {
+                event_id,
+                is_falling_back,
+            }) => Relation::Thread(Thread { event_id, in_reply_to, is_falling_back }),
+        }
+    } else if let Some(in_reply_to) = in_reply_to {
+        Relation::Reply { in_reply_to }
+    } else if let Some(RelationDeHelper::Unknown(c)) = relation {
+        Relation::_Custom(c)
+    } else {
+        return Err(de::Error::missing_field("m.in_reply_to or rel_type"));
+    };
+
+    Ok(Some(rel))
 }
 
 impl<C> Serialize for Relation<C>
@@ -66,133 +67,157 @@ where
 {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
-        S: Serializer,
+        S: serde::Serializer,
     {
-        let json_repr = match self {
-            Relation::Reply { in_reply_to } => {
-                EventWithRelatesToJsonRepr::<C>::new(RelatesToJsonRepr {
-                    in_reply_to: Some(in_reply_to.clone()),
-                    ..Default::default()
-                })
-            }
-            Relation::Replacement(Replacement { event_id, new_content }) => {
-                EventWithRelatesToJsonRepr {
-                    relates_to: RelatesToJsonRepr {
-                        relation: Some(RelationJsonRepr::Replacement(ReplacementJsonRepr {
-                            event_id: event_id.clone(),
-                        })),
-                        ..Default::default()
-                    },
-                    new_content: Some(new_content.clone()),
-                }
-            }
-            Relation::Thread(Thread { event_id, in_reply_to, is_falling_back }) => {
-                EventWithRelatesToJsonRepr::new(RelatesToJsonRepr {
-                    in_reply_to: in_reply_to.clone(),
-                    relation: Some(RelationJsonRepr::ThreadStable(ThreadStableJsonRepr {
-                        event_id: event_id.clone(),
-                        is_falling_back: *is_falling_back,
-                    })),
-                })
-            }
-            Relation::_Custom => EventWithRelatesToJsonRepr::<C>::default(),
-        };
+        let (relates_to, new_content) = self.clone().into_parts();
 
-        json_repr.serialize(serializer)
+        EventWithRelatesToSerHelper { relates_to, new_content }.serialize(serializer)
     }
 }
 
+#[derive(Deserialize)]
+struct EventWithRelatesToDeHelper<C> {
+    #[serde(rename = "m.relates_to")]
+    relates_to: Option<RelatesToDeHelper>,
+
+    #[serde(rename = "m.new_content")]
+    new_content: Option<C>,
+}
+
+#[derive(Deserialize)]
+struct RelatesToDeHelper {
+    #[serde(rename = "m.in_reply_to")]
+    in_reply_to: Option<InReplyTo>,
+
+    #[serde(flatten)]
+    relation: Option<RelationDeHelper>,
+}
+
+#[derive(Deserialize)]
+#[serde(untagged)]
+enum RelationDeHelper {
+    Known(KnownRelationDeHelper),
+    Unknown(CustomRelation),
+}
+
+#[derive(Deserialize)]
+#[serde(tag = "rel_type")]
+enum KnownRelationDeHelper {
+    #[serde(rename = "m.replace")]
+    Replacement(ReplacementJsonRepr),
+
+    #[serde(rename = "m.thread")]
+    Thread(ThreadDeHelper),
+
+    #[serde(rename = "io.element.thread")]
+    ThreadUnstable(ThreadUnstableDeHelper),
+}
+
+/// A replacement relation without `m.new_content`.
 #[derive(Deserialize, Serialize)]
-struct EventWithRelatesToJsonRepr<C> {
-    #[serde(rename = "m.relates_to", default, skip_serializing_if = "RelatesToJsonRepr::is_empty")]
-    relates_to: RelatesToJsonRepr,
+pub(super) struct ReplacementJsonRepr {
+    event_id: OwnedEventId,
+}
+
+/// A thread relation without the reply fallback, with stable names.
+#[derive(Deserialize)]
+struct ThreadDeHelper {
+    event_id: OwnedEventId,
+
+    #[serde(default)]
+    is_falling_back: bool,
+}
+
+/// A thread relation without the reply fallback, with unstable names.
+#[derive(Deserialize)]
+struct ThreadUnstableDeHelper {
+    event_id: OwnedEventId,
+
+    #[serde(rename = "io.element.show_reply", default)]
+    is_falling_back: bool,
+}
+
+#[derive(Serialize)]
+pub(super) struct EventWithRelatesToSerHelper<C> {
+    #[serde(rename = "m.relates_to")]
+    relates_to: RelationSerHelper,
 
     #[serde(rename = "m.new_content", skip_serializing_if = "Option::is_none")]
     new_content: Option<C>,
 }
 
-impl<C> EventWithRelatesToJsonRepr<C> {
-    fn new(relates_to: RelatesToJsonRepr) -> Self {
-        Self { relates_to, new_content: None }
-    }
-}
-
-impl<C> Default for EventWithRelatesToJsonRepr<C> {
-    fn default() -> Self {
-        Self { relates_to: RelatesToJsonRepr::default(), new_content: None }
-    }
-}
-
-/// Struct modeling the different ways relationships can be expressed in a `m.relates_to` field of
-/// an event.
-#[derive(Default, Deserialize, Serialize)]
-struct RelatesToJsonRepr {
-    #[serde(rename = "m.in_reply_to", skip_serializing_if = "Option::is_none")]
-    in_reply_to: Option<InReplyTo>,
-
-    #[serde(flatten, skip_serializing_if = "Option::is_none")]
-    relation: Option<RelationJsonRepr>,
-}
-
-impl RelatesToJsonRepr {
-    fn is_empty(&self) -> bool {
-        self.in_reply_to.is_none() && self.relation.is_none()
-    }
-}
-
 /// A relation, which associates new information to an existing event.
-#[derive(Clone, Deserialize, Serialize)]
+#[derive(Serialize)]
 #[serde(tag = "rel_type")]
-enum RelationJsonRepr {
+pub(super) enum RelationSerHelper {
     /// An event that replaces another event.
     #[serde(rename = "m.replace")]
     Replacement(ReplacementJsonRepr),
 
     /// An event that belongs to a thread, with stable names.
     #[serde(rename = "m.thread")]
-    ThreadStable(ThreadStableJsonRepr),
-
-    /// An event that belongs to a thread, with unstable names.
-    #[serde(rename = "io.element.thread")]
-    ThreadUnstable(ThreadUnstableJsonRepr),
+    Thread(Thread),
 
     /// An unknown relation type.
-    ///
-    /// Not available in the public API, but exists here so deserialization
-    /// doesn't fail with new / custom `rel_type`s.
-    #[serde(other)]
-    Unknown,
+    #[serde(untagged)]
+    Custom(CustomSerHelper),
 }
 
-#[derive(Clone, Deserialize, Serialize)]
-struct ReplacementJsonRepr {
-    event_id: OwnedEventId,
+impl<C> Relation<C> {
+    fn into_parts(self) -> (RelationSerHelper, Option<C>) {
+        match self {
+            Relation::Replacement(Replacement { event_id, new_content }) => (
+                RelationSerHelper::Replacement(ReplacementJsonRepr { event_id }),
+                Some(new_content),
+            ),
+            Relation::Reply { in_reply_to } => {
+                (RelationSerHelper::Custom(in_reply_to.into()), None)
+            }
+            Relation::Thread(t) => (RelationSerHelper::Thread(t), None),
+            Relation::_Custom(c) => (RelationSerHelper::Custom(c.into()), None),
+        }
+    }
+
+    pub(super) fn serialize_data(&self) -> JsonObject
+    where
+        C: Clone,
+    {
+        let (relates_to, _) = self.clone().into_parts();
+
+        match serde_json::to_value(relates_to).expect("relation serialization to succeed") {
+            JsonValue::Object(mut obj) => {
+                obj.remove("rel_type");
+                obj
+            }
+            _ => panic!("all relations must serialize to objects"),
+        }
+    }
 }
 
-/// A thread relation without the reply fallback, with stable names.
-#[derive(Clone, Deserialize, Serialize)]
-struct ThreadStableJsonRepr {
-    /// The ID of the root message in the thread.
-    event_id: OwnedEventId,
+#[derive(Default, Serialize)]
+pub(super) struct CustomSerHelper {
+    #[serde(rename = "m.in_reply_to", skip_serializing_if = "Option::is_none")]
+    in_reply_to: Option<InReplyTo>,
 
-    /// Whether the `m.in_reply_to` field is a fallback for older clients or a real reply in a
-    /// thread.
-    #[serde(default, skip_serializing_if = "ruma_common::serde::is_default")]
-    is_falling_back: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    rel_type: Option<String>,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
+    event_id: Option<OwnedEventId>,
+
+    #[serde(flatten, skip_serializing_if = "JsonObject::is_empty")]
+    data: JsonObject,
 }
 
-/// A thread relation without the reply fallback, with unstable names.
-#[derive(Clone, Deserialize, Serialize)]
-struct ThreadUnstableJsonRepr {
-    /// The ID of the root message in the thread.
-    event_id: OwnedEventId,
+impl From<InReplyTo> for CustomSerHelper {
+    fn from(value: InReplyTo) -> Self {
+        Self { in_reply_to: Some(value), ..Default::default() }
+    }
+}
 
-    /// Whether the `m.in_reply_to` field is a fallback for older clients or a real reply in a
-    /// thread.
-    #[serde(
-        rename = "io.element.show_reply",
-        default,
-        skip_serializing_if = "ruma_common::serde::is_default"
-    )]
-    is_falling_back: bool,
+impl From<CustomRelation> for CustomSerHelper {
+    fn from(value: CustomRelation) -> Self {
+        let CustomRelation { rel_type, event_id, data } = value;
+        Self { rel_type: Some(rel_type), event_id: Some(event_id), data, ..Default::default() }
+    }
 }