+++ crates/ruma-common/src/lib.rs
@@ -33,6 +33,7 @@ pub mod presence;
 pub mod push;
 pub mod room;
 pub mod serde;
+pub mod space;
 pub mod thirdparty;
 mod time;
 pub mod to_device;